Исправление основных ошибок в ходе сценария работы

+++ common/src/main/java/ru/liga/Common/dtos/ShortOrderItemDTO.java
@@ -17,5 +17,5 @@ public class ShortOrderItemDTO {
 
     @Schema(description = "Количество")
     @JsonProperty("quantity")
-    private byte quantity;
+    private int quantity;
 }
