+++ common/src/main/java/ru/liga/common/dtos/OrderItemDTO.java
@@ -21,7 +21,7 @@ public class OrderItemDTO {
 
     @Schema(description = "Количество")
     @JsonProperty("quantity")
-    private byte quantity;
+    private int quantity;
 
     @Schema(description = "Описание позиции")
     @JsonProperty("description")
